+++ frontend/dialob-composer-material/src/dialob/index.ts
@@ -1,6 +1,3 @@
-// Public export from dialob composer core
-
 export { useComposer } from "./react/useComposer"
 export { ComposerContext, ComposerProvider} from './react/ComposerContext';
 export * from './types';
-