+++ frontend/dialob-composer-material/src/items/Group.tsx
@@ -1,0 +1,12 @@
+import { Typography } from '@mui/material';
+import React from 'react';
+
+const Group: React.FC = () => {
+  return (
+    <>
+      <Typography>Group</Typography>
+    </>
+  );
+};
+
+export { Group };