+++ frontend/dialob-composer-material/src/items/index.ts
@@ -1,0 +1,11 @@
+import { Group } from "./Group";
+import { SimpleField } from "./SimpleField";
+import { Note } from "./Note";
+import { itemFactory } from "./ItemFactory";
+
+export default {
+  Group,
+  SimpleField,
+  Note,
+  itemFactory
+};