+++ frontend/dialob-composer-material/src/theme/siteTheme.ts
@@ -270,6 +270,7 @@ const siteTheme = createTheme({
         root: {
           fontVariant: 'body2',
           borderRadius: 0,
+          borderColor: palette.text.hint,
           textTransform: 'capitalize',
           borderWidth: '2px solid !important',
         }
@@ -307,6 +308,14 @@ const siteTheme = createTheme({
         },
       }
     },
+
+    MuiTableCell: {
+      styleOverrides: {
+        root: {
+          padding: 0,
+        },
+      }
+    }
 
   },